Refactor ErrorHandlingContainerModal.tsx & replace copy-to-clipboard with navigator.clipboard
