+++ db-viewer-ui/package.json
@@ -49,7 +49,6 @@
   "dependencies": {
     "@skbkontur/react-stack-layout": "^1.0.3",
     "@skbkontur/react-ui-validations": "^1.8.3",
-    "copy-to-clipboard": "^3.3.1",
     "date-fns": "^2.29.1",
     "lodash": "^4.17.20",
     "tslib": "^2.3.0"
